Revise link resolution.
relLangURL resolves a path relative to the whole website(baseURL), which
makes it unsuitable for links relative to the current page. relref
could be used, but it doesn't correctly handle external links.

It appears that the convention in Hugo is to leave links rendered from
markdown as-is, and use the ref/relref shortcodes when linking to other
pages on the website. Shadocs goes against this convention by parsing
and handling links rendered by markdown. Despite this, it is rather
convenient. However, it uses relLangURL to resolve links, which is
naively relative to the baseURL, making it unsuitable for relative
paths. relref could be used, but while relref is smarter, it doesn't
work with external links. Neither can be used on their own.

To resolve this, the URL is parsed. If it has a scheme, then relLangURL
is used as before. Otherwise, relref is used.